did the pulsar

+++ wallet-management/src/main/java/africa/semicolon/infrastructure/adapter/monnify/dtos/request/InitializeTransferRequest.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 public class InitializeTransferRequest {
 
-    private Float amount;
+    private double amount;
     private String reference;
     private String narration;
     private String receiverAccountNumber;
